MR-Client refactoring/restructuring
