--- packages/mr-client/components/models/Labyrinth/Labyrinth.stories.mdx
@@ -1,29 +1,0 @@
-import { Loader, OrbitControls } from '@react-three/drei'
-import { Canvas } from '@react-three/fiber'
-import { Meta, Story, Preview, Props } from '@storybook/addon-docs'
-import { Suspense } from 'react'
-import Lighting from '../../scenes/MainGameScene/components/Lighting'
-import Labyrinth from './Labyrinth'
-
-<Meta title='Models/Labyrinth' />
-
-# Labyrinth
-
-This is the basic labyrinth component with a basic lighting setup.
-
-<Preview>
-  <Story name='Labyrinth'>
-    <Canvas style={{ height: '100vh', backgroundColor: 'black' }}>
-      <Suspense fallback={null}>
-        <Labyrinth />
-      </Suspense>
-      <Lighting />
-      <Loader />
-      <OrbitControls />
-    </Canvas>
-  </Story>
-</Preview>
-
-## Props
-
-<Props of={Labyrinth} />
