+++ packages/mr-client/src/Player/Player.stories.mdx
@@ -1,0 +1,26 @@
+import { Loader, OrbitControls } from '@react-three/drei'
+import { Canvas } from '@react-three/fiber'
+import { Meta, Story, Preview, Props } from '@storybook/addon-docs'
+import Lighting from '../Lighting/Lighting'
+import Player from './Player'
+
+<Meta title='Player' />
+
+# Player
+
+This is the primitive MVP version of the Player.
+
+<Preview>
+  <Story name='Player'>
+    <Canvas style={{ height: '100vh', backgroundColor: 'black' }}>
+      <Player />
+      <Lighting />
+      <Loader />
+      <OrbitControls />
+    </Canvas>
+  </Story>
+</Preview>
+
+## Props
+
+<Props of={Player} />
