+++ packages/mr-client/src/Player/Player.tsx
@@ -2,7 +2,7 @@ export default function Player(): JSX.Element {
   return (
     <mesh castShadow receiveShadow>
       <sphereGeometry attach='geometry' args={[0.5, 32, 200]} />
-      <meshStandardMaterial color='red' />
+      <meshStandardMaterial color='lightgreen' />
     </mesh>
   )
 }
